+++ shoop/admin/modules/contact_groups/__init__.py
@@ -10,7 +10,9 @@ from __future__ import unicode_literals
 from django.utils.translation import ugettext_lazy as _
 
 from shoop.admin.base import AdminModule, MenuEntry
-from shoop.admin.utils.urls import derive_model_url, get_edit_and_list_urls
+from shoop.admin.utils.urls import (
+    admin_url, derive_model_url, get_edit_and_list_urls
+)
 from shoop.core.models import ContactGroup
 
 
@@ -20,7 +22,13 @@ class ContactGroupModule(AdminModule):
     breadcrumbs_menu_entry = MenuEntry(name, url="shoop_admin:contact-group.list")
 
     def get_urls(self):
-        return get_edit_and_list_urls(
+        return [
+            admin_url(
+                "^contact-group/(?P<pk>\d+)/delete/$",
+                "shoop.admin.modules.contact_groups.views.ContactGroupDeleteView",
+                name="contact-group.delete"
+            )
+        ] + get_edit_and_list_urls(
             url_prefix="^contact-groups",
             view_template="shoop.admin.modules.contact_groups.views.ContactGroup%sView",
             name_template="contact-group.%s"